feat: TimerInfoStackView 구현
