+++ Clock/Clock/Presentation/Scene/Timer/CreateTimer/Subview/TimerInfoStackView.swift
@@ -1,0 +1,75 @@
+//
+//  TimerInfoStackView.swift
+//  Clock
+//
+//  Created by 이수현 on 5/21/25.
+//
+
+import UIKit
+
+final class TimerInfoStackView: UIStackView {
+    private let labelStackView: UIStackView = {
+        let stackView = UIStackView()
+        stackView.backgroundColor = .gray.withAlphaComponent(0.2)
+        stackView.axis = .horizontal
+        stackView.layoutMargins = .init(top: 0, left: 12, bottom: 0, right: 12)
+        stackView.isLayoutMarginsRelativeArrangement = true
+        stackView.spacing = 8
+        return stackView
+    }()
+
+    private let labelLabel: UILabel = {
+        let label = UILabel()
+        label.text = "레이블"
+        label.textColor = .white
+        label.font = .systemFont(ofSize: 16)
+        label.sizeToFit()
+        return label
+    }()
+
+    private let labelTextField: UITextField = {
+        let textField = UITextField()
+        textField.placeholder = "타이머"
+        textField.font = .systemFont(ofSize: 16)
+        textField.textAlignment = .right
+        textField.clearButtonMode = .whileEditing
+        return textField
+    }()
+
+    private let soundButton = SoundButton()
+
+    override init(frame: CGRect) {
+        super.init(frame: frame)
+
+        setAttributes()
+        setHierarchy()
+    }
+
+    @available(*, unavailable)
+    required init(coder: NSCoder) {
+        fatalError()
+    }
+}
+
+private extension TimerInfoStackView {
+    func setAttributes() {
+        self.backgroundColor = .background
+        self.axis = .vertical
+        self.spacing = 1
+        self.distribution = .fillEqually
+        self.layer.cornerRadius = 12
+        self.clipsToBounds = true
+    }
+
+    func setHierarchy() {
+        [
+            labelLabel,
+            labelTextField
+        ].forEach { labelStackView.addArrangedSubview($0) }
+
+        [
+            labelStackView,
+            soundButton
+        ].forEach { self.addArrangedSubview($0) }
+    }
+}
